requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
node==0.9.28
npm==0.1.1
odict==1.8.0
optional-django==0.1.0
plumber==1.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
Pillow==6.1.0
